eslint separate enum
