+++ frontend/src/components/SimpleTests/TeacherOrStudentEnum.tsx
@@ -1,0 +1,4 @@
+export enum TeacherOrStudent {
+    TEACHER,
+    STUDENT,
+}
